Implement player shooting bullets.
All it does is create the bullets, and they move.
Board handles the actual making of them, since there's a lot of moving parts.
Player just calls doShoot, AbstractThing just checks if blocked and defers to Board.
Board garbageCollects the bullets at the end of a cycle (so no dangling pointer issues).