Implement :touch and #play
Because I like when things react to my touching them.
Put in place the label seeking code. I'm being careful to have the
program specified as a parameter, since the #bind issue is looming
over my head.

The scroll in TOWN.ZZT should now play music when you touch it.